pyproject.toml
[project]
name = "research-form-submission"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-fasthtml>=0.6.4",
    "requests>=2.32.3",
    "toml>=0.10.2",
]
